Print Summary Diagram of Multiwave Object

Takes a multiwave object as input and plots a diagram of its structure in the plotting window using grViz() from the DiagrammeR package. Red boxes indicate slots that have not yet been filled, blue boxes indicate that the slot is filled.

Usage

multiwave_diagram(x, height = NULL, width = NULL)

Arguments

x

An object of class multiwave.

height

The height in pixels of the diagram. Defaults to NULL , which produces default height.

width

The width in pixels of the diagram. Defaults to NULL, which produces the default width.

Value

Returns an object of class htmlwidget

displaying the structure of the x.

Examples

MySurvey <- multiwave(phases = 2, waves = c(1, 3))
multiwave_diagram(MySurvey)