Select Sampling Units based on Stratified Random Sampling

Requires two dataframes or matrices: data with a column strata which specifies stratum membership for each unit in the population and a second dataframe design_data with one row per strata level with a column design_strata that indicates the unique levels of strata in data and n_allocated that specifies the number to be sampled from each stratum. sample_strata selects the units to sample by selecting a random sample of the desired size within each stratum. The second dataframe can be the output of allocate_wave() or optimum_allocation().

Usage

sample_strata(
  data,
  strata,
  id,
  already_sampled = NULL,
  design_data,
  design_strata = "strata",
  n_allocated = "n_to_sample",
  probs = NULL,
  wave = NULL,
  warn_prob_overwrite = TRUE
)

Arguments

data

A data frame or matrix with one row for each sampling unit in the population, one column specifying each unit's stratum, and one column with a unique identifier for each unit.

strata

a character string specifying the name of column in data which indicates stratum membership.

id

a character string specifying the name of the column in data that uniquely identifies each unit.

already_sampled

a character sting specifying the name of the column in data which indicates (1/0 or Y/N) whether a unit has already been sampled in a prior wave. Defaults to NULL which means that none have been sampled yet.

design_data

a dataframe or matrix with one row for each stratum that subdivides the population, one column specifying the stratum name, and one column indicating the number of samples allocated to each stratum.

design_strata

a character string specifying the name of the column in design_data that contains the stratum levels. Defaults to "strata".

n_allocated

a character string specifying the name of the column in design_data that indicates the n allocated to each stratum. Defaults to "n_to_sample".

probs

a character string specifying the name of the column in in design_data that indicates the sampling probability for each stratum, or a formula indicating how the sampling probabilities can be computed. From existing columns. If specified, a new column containing the sampling probability attached to each sampled unit will be created in the outputted dataframe. This column will be named "sampling_prob". Defaults to NULL.

wave

A numeric value or character string indicating the sampling wave. If specified, the input is appended to "sample_indicator" in the new the sample indicator column name (as long as such columns name do not already exist in data). Defaults to NULL. This argument does not apply when sample_strata() is called inside allocate_wave().

warn_prob_overwrite

Logical indicator for whether warning should be printed if probs is specified and a "sampling_prob" columns is going to be overwritten. Defaults to TRUE. If function is called inside apply_multiwave(), then defaults to FALSE

Value

returns data as a dataframe with a new column named "sample_indicator" containing a binary (1/0) indicator of whether each unit should be sampled. If wave argument is specified, then the given input is appended to the name "sample_indicator". If probs argument is specified, then the dataframe will also contain a new column named "sampling_prob" holding the sampling probabilities for each sampled element.

Examples

# Define a design dataframe
design <- data.frame(
  strata = c("setosa", "virginica", "versicolor"),
  npop = c(50, 50, 50),
  n_to_sample = c(5, 5, 5)
)

# Make sure there is an id column
iris$id <- 1:nrow(iris)

# Run
sample_strata(
  data = iris, strata = "Species", id = "id",
  design_data = design, design_strata = "strata",
  n_allocated = "n_to_sample"
)
#>     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species  id
#> 1            5.1         3.5          1.4         0.2     setosa   1
#> 2            4.9         3.0          1.4         0.2     setosa   2
#> 3            4.7         3.2          1.3         0.2     setosa   3
#> 4            4.6         3.1          1.5         0.2     setosa   4
#> 5            5.0         3.6          1.4         0.2     setosa   5
#> 6            5.4         3.9          1.7         0.4     setosa   6
#> 7            4.6         3.4          1.4         0.3     setosa   7
#> 8            5.0         3.4          1.5         0.2     setosa   8
#> 9            4.4         2.9          1.4         0.2     setosa   9
#> 10           4.9         3.1          1.5         0.1     setosa  10
#> 11           5.4         3.7          1.5         0.2     setosa  11
#> 12           4.8         3.4          1.6         0.2     setosa  12
#> 13           4.8         3.0          1.4         0.1     setosa  13
#> 14           4.3         3.0          1.1         0.1     setosa  14
#> 15           5.8         4.0          1.2         0.2     setosa  15
#> 16           5.7         4.4          1.5         0.4     setosa  16
#> 17           5.4         3.9          1.3         0.4     setosa  17
#> 18           5.1         3.5          1.4         0.3     setosa  18
#> 19           5.7         3.8          1.7         0.3     setosa  19
#> 20           5.1         3.8          1.5         0.3     setosa  20
#> 21           5.4         3.4          1.7         0.2     setosa  21
#> 22           5.1         3.7          1.5         0.4     setosa  22
#> 23           4.6         3.6          1.0         0.2     setosa  23
#> 24           5.1         3.3          1.7         0.5     setosa  24
#> 25           4.8         3.4          1.9         0.2     setosa  25
#> 26           5.0         3.0          1.6         0.2     setosa  26
#> 27           5.0         3.4          1.6         0.4     setosa  27
#> 28           5.2         3.5          1.5         0.2     setosa  28
#> 29           5.2         3.4          1.4         0.2     setosa  29
#> 30           4.7         3.2          1.6         0.2     setosa  30
#> 31           4.8         3.1          1.6         0.2     setosa  31
#> 32           5.4         3.4          1.5         0.4     setosa  32
#> 33           5.2         4.1          1.5         0.1     setosa  33
#> 34           5.5         4.2          1.4         0.2     setosa  34
#> 35           4.9         3.1          1.5         0.2     setosa  35
#> 36           5.0         3.2          1.2         0.2     setosa  36
#> 37           5.5         3.5          1.3         0.2     setosa  37
#> 38           4.9         3.6          1.4         0.1     setosa  38
#> 39           4.4         3.0          1.3         0.2     setosa  39
#> 40           5.1         3.4          1.5         0.2     setosa  40
#> 41           5.0         3.5          1.3         0.3     setosa  41
#> 42           4.5         2.3          1.3         0.3     setosa  42
#> 43           4.4         3.2          1.3         0.2     setosa  43
#> 44           5.0         3.5          1.6         0.6     setosa  44
#> 45           5.1         3.8          1.9         0.4     setosa  45
#> 46           4.8         3.0          1.4         0.3     setosa  46
#> 47           5.1         3.8          1.6         0.2     setosa  47
#> 48           4.6         3.2          1.4         0.2     setosa  48
#> 49           5.3         3.7          1.5         0.2     setosa  49
#> 50           5.0         3.3          1.4         0.2     setosa  50
#> 51           7.0         3.2          4.7         1.4 versicolor  51
#> 52           6.4         3.2          4.5         1.5 versicolor  52
#> 53           6.9         3.1          4.9         1.5 versicolor  53
#> 54           5.5         2.3          4.0         1.3 versicolor  54
#> 55           6.5         2.8          4.6         1.5 versicolor  55
#> 56           5.7         2.8          4.5         1.3 versicolor  56
#> 57           6.3         3.3          4.7         1.6 versicolor  57
#> 58           4.9         2.4          3.3         1.0 versicolor  58
#> 59           6.6         2.9          4.6         1.3 versicolor  59
#> 60           5.2         2.7          3.9         1.4 versicolor  60
#> 61           5.0         2.0          3.5         1.0 versicolor  61
#> 62           5.9         3.0          4.2         1.5 versicolor  62
#> 63           6.0         2.2          4.0         1.0 versicolor  63
#> 64           6.1         2.9          4.7         1.4 versicolor  64
#> 65           5.6         2.9          3.6         1.3 versicolor  65
#> 66           6.7         3.1          4.4         1.4 versicolor  66
#> 67           5.6         3.0          4.5         1.5 versicolor  67
#> 68           5.8         2.7          4.1         1.0 versicolor  68
#> 69           6.2         2.2          4.5         1.5 versicolor  69
#> 70           5.6         2.5          3.9         1.1 versicolor  70
#> 71           5.9         3.2          4.8         1.8 versicolor  71
#> 72           6.1         2.8          4.0         1.3 versicolor  72
#> 73           6.3         2.5          4.9         1.5 versicolor  73
#> 74           6.1         2.8          4.7         1.2 versicolor  74
#> 75           6.4         2.9          4.3         1.3 versicolor  75
#> 76           6.6         3.0          4.4         1.4 versicolor  76
#> 77           6.8         2.8          4.8         1.4 versicolor  77
#> 78           6.7         3.0          5.0         1.7 versicolor  78
#> 79           6.0         2.9          4.5         1.5 versicolor  79
#> 80           5.7         2.6          3.5         1.0 versicolor  80
#> 81           5.5         2.4          3.8         1.1 versicolor  81
#> 82           5.5         2.4          3.7         1.0 versicolor  82
#> 83           5.8         2.7          3.9         1.2 versicolor  83
#> 84           6.0         2.7          5.1         1.6 versicolor  84
#> 85           5.4         3.0          4.5         1.5 versicolor  85
#> 86           6.0         3.4          4.5         1.6 versicolor  86
#> 87           6.7         3.1          4.7         1.5 versicolor  87
#> 88           6.3         2.3          4.4         1.3 versicolor  88
#> 89           5.6         3.0          4.1         1.3 versicolor  89
#> 90           5.5         2.5          4.0         1.3 versicolor  90
#> 91           5.5         2.6          4.4         1.2 versicolor  91
#> 92           6.1         3.0          4.6         1.4 versicolor  92
#> 93           5.8         2.6          4.0         1.2 versicolor  93
#> 94           5.0         2.3          3.3         1.0 versicolor  94
#> 95           5.6         2.7          4.2         1.3 versicolor  95
#> 96           5.7         3.0          4.2         1.2 versicolor  96
#> 97           5.7         2.9          4.2         1.3 versicolor  97
#> 98           6.2         2.9          4.3         1.3 versicolor  98
#> 99           5.1         2.5          3.0         1.1 versicolor  99
#> 100          5.7         2.8          4.1         1.3 versicolor 100
#> 101          6.3         3.3          6.0         2.5  virginica 101
#> 102          5.8         2.7          5.1         1.9  virginica 102
#> 103          7.1         3.0          5.9         2.1  virginica 103
#> 104          6.3         2.9          5.6         1.8  virginica 104
#> 105          6.5         3.0          5.8         2.2  virginica 105
#> 106          7.6         3.0          6.6         2.1  virginica 106
#> 107          4.9         2.5          4.5         1.7  virginica 107
#> 108          7.3         2.9          6.3         1.8  virginica 108
#> 109          6.7         2.5          5.8         1.8  virginica 109
#> 110          7.2         3.6          6.1         2.5  virginica 110
#> 111          6.5         3.2          5.1         2.0  virginica 111
#> 112          6.4         2.7          5.3         1.9  virginica 112
#> 113          6.8         3.0          5.5         2.1  virginica 113
#> 114          5.7         2.5          5.0         2.0  virginica 114
#> 115          5.8         2.8          5.1         2.4  virginica 115
#> 116          6.4         3.2          5.3         2.3  virginica 116
#> 117          6.5         3.0          5.5         1.8  virginica 117
#> 118          7.7         3.8          6.7         2.2  virginica 118
#> 119          7.7         2.6          6.9         2.3  virginica 119
#> 120          6.0         2.2          5.0         1.5  virginica 120
#> 121          6.9         3.2          5.7         2.3  virginica 121
#> 122          5.6         2.8          4.9         2.0  virginica 122
#> 123          7.7         2.8          6.7         2.0  virginica 123
#> 124          6.3         2.7          4.9         1.8  virginica 124
#> 125          6.7         3.3          5.7         2.1  virginica 125
#> 126          7.2         3.2          6.0         1.8  virginica 126
#> 127          6.2         2.8          4.8         1.8  virginica 127
#> 128          6.1         3.0          4.9         1.8  virginica 128
#> 129          6.4         2.8          5.6         2.1  virginica 129
#> 130          7.2         3.0          5.8         1.6  virginica 130
#> 131          7.4         2.8          6.1         1.9  virginica 131
#> 132          7.9         3.8          6.4         2.0  virginica 132
#> 133          6.4         2.8          5.6         2.2  virginica 133
#> 134          6.3         2.8          5.1         1.5  virginica 134
#> 135          6.1         2.6          5.6         1.4  virginica 135
#> 136          7.7         3.0          6.1         2.3  virginica 136
#> 137          6.3         3.4          5.6         2.4  virginica 137
#> 138          6.4         3.1          5.5         1.8  virginica 138
#> 139          6.0         3.0          4.8         1.8  virginica 139
#> 140          6.9         3.1          5.4         2.1  virginica 140
#> 141          6.7         3.1          5.6         2.4  virginica 141
#> 142          6.9         3.1          5.1         2.3  virginica 142
#> 143          5.8         2.7          5.1         1.9  virginica 143
#> 144          6.8         3.2          5.9         2.3  virginica 144
#> 145          6.7         3.3          5.7         2.5  virginica 145
#> 146          6.7         3.0          5.2         2.3  virginica 146
#> 147          6.3         2.5          5.0         1.9  virginica 147
#> 148          6.5         3.0          5.2         2.0  virginica 148
#> 149          6.2         3.4          5.4         2.3  virginica 149
#> 150          5.9         3.0          5.1         1.8  virginica 150
#>     sample_indicator
#> 1                  0
#> 2                  0
#> 3                  0
#> 4                  0
#> 5                  0
#> 6                  0
#> 7                  0
#> 8                  1
#> 9                  0
#> 10                 1
#> 11                 0
#> 12                 0
#> 13                 0
#> 14                 0
#> 15                 0
#> 16                 0
#> 17                 0
#> 18                 0
#> 19                 0
#> 20                 0
#> 21                 0
#> 22                 0
#> 23                 0
#> 24                 1
#> 25                 0
#> 26                 0
#> 27                 0
#> 28                 0
#> 29                 1
#> 30                 0
#> 31                 0
#> 32                 0
#> 33                 0
#> 34                 0
#> 35                 0
#> 36                 0
#> 37                 0
#> 38                 0
#> 39                 0
#> 40                 0
#> 41                 0
#> 42                 0
#> 43                 0
#> 44                 1
#> 45                 0
#> 46                 0
#> 47                 0
#> 48                 0
#> 49                 0
#> 50                 0
#> 51                 0
#> 52                 0
#> 53                 0
#> 54                 0
#> 55                 0
#> 56                 0
#> 57                 0
#> 58                 0
#> 59                 1
#> 60                 0
#> 61                 0
#> 62                 0
#> 63                 0
#> 64                 0
#> 65                 0
#> 66                 0
#> 67                 0
#> 68                 0
#> 69                 0
#> 70                 0
#> 71                 0
#> 72                 0
#> 73                 1
#> 74                 0
#> 75                 0
#> 76                 0
#> 77                 0
#> 78                 0
#> 79                 0
#> 80                 0
#> 81                 0
#> 82                 0
#> 83                 0
#> 84                 1
#> 85                 0
#> 86                 0
#> 87                 0
#> 88                 0
#> 89                 0
#> 90                 0
#> 91                 0
#> 92                 1
#> 93                 0
#> 94                 0
#> 95                 0
#> 96                 0
#> 97                 0
#> 98                 1
#> 99                 0
#> 100                0
#> 101                0
#> 102                0
#> 103                0
#> 104                0
#> 105                0
#> 106                0
#> 107                1
#> 108                0
#> 109                0
#> 110                0
#> 111                0
#> 112                0
#> 113                0
#> 114                0
#> 115                0
#> 116                0
#> 117                0
#> 118                0
#> 119                1
#> 120                0
#> 121                0
#> 122                0
#> 123                0
#> 124                0
#> 125                0
#> 126                0
#> 127                0
#> 128                1
#> 129                0
#> 130                0
#> 131                0
#> 132                0
#> 133                1
#> 134                0
#> 135                1
#> 136                0
#> 137                0
#> 138                0
#> 139                0
#> 140                0
#> 141                0
#> 142                0
#> 143                0
#> 144                0
#> 145                0
#> 146                0
#> 147                0
#> 148                0
#> 149                0
#> 150                0

# To include probs as a formula
sample_strata(
  data = iris, strata = "Species", id = "id",
  design_data = design, design_strata = "strata",
  n_allocated = "n_to_sample", probs = ~n_to_sample/npop
)
#>     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species  id
#> 1            5.1         3.5          1.4         0.2     setosa   1
#> 2            4.9         3.0          1.4         0.2     setosa   2
#> 3            4.7         3.2          1.3         0.2     setosa   3
#> 4            4.6         3.1          1.5         0.2     setosa   4
#> 5            5.0         3.6          1.4         0.2     setosa   5
#> 6            5.4         3.9          1.7         0.4     setosa   6
#> 7            4.6         3.4          1.4         0.3     setosa   7
#> 8            5.0         3.4          1.5         0.2     setosa   8
#> 9            4.4         2.9          1.4         0.2     setosa   9
#> 10           4.9         3.1          1.5         0.1     setosa  10
#> 11           5.4         3.7          1.5         0.2     setosa  11
#> 12           4.8         3.4          1.6         0.2     setosa  12
#> 13           4.8         3.0          1.4         0.1     setosa  13
#> 14           4.3         3.0          1.1         0.1     setosa  14
#> 15           5.8         4.0          1.2         0.2     setosa  15
#> 16           5.7         4.4          1.5         0.4     setosa  16
#> 17           5.4         3.9          1.3         0.4     setosa  17
#> 18           5.1         3.5          1.4         0.3     setosa  18
#> 19           5.7         3.8          1.7         0.3     setosa  19
#> 20           5.1         3.8          1.5         0.3     setosa  20
#> 21           5.4         3.4          1.7         0.2     setosa  21
#> 22           5.1         3.7          1.5         0.4     setosa  22
#> 23           4.6         3.6          1.0         0.2     setosa  23
#> 24           5.1         3.3          1.7         0.5     setosa  24
#> 25           4.8         3.4          1.9         0.2     setosa  25
#> 26           5.0         3.0          1.6         0.2     setosa  26
#> 27           5.0         3.4          1.6         0.4     setosa  27
#> 28           5.2         3.5          1.5         0.2     setosa  28
#> 29           5.2         3.4          1.4         0.2     setosa  29
#> 30           4.7         3.2          1.6         0.2     setosa  30
#> 31           4.8         3.1          1.6         0.2     setosa  31
#> 32           5.4         3.4          1.5         0.4     setosa  32
#> 33           5.2         4.1          1.5         0.1     setosa  33
#> 34           5.5         4.2          1.4         0.2     setosa  34
#> 35           4.9         3.1          1.5         0.2     setosa  35
#> 36           5.0         3.2          1.2         0.2     setosa  36
#> 37           5.5         3.5          1.3         0.2     setosa  37
#> 38           4.9         3.6          1.4         0.1     setosa  38
#> 39           4.4         3.0          1.3         0.2     setosa  39
#> 40           5.1         3.4          1.5         0.2     setosa  40
#> 41           5.0         3.5          1.3         0.3     setosa  41
#> 42           4.5         2.3          1.3         0.3     setosa  42
#> 43           4.4         3.2          1.3         0.2     setosa  43
#> 44           5.0         3.5          1.6         0.6     setosa  44
#> 45           5.1         3.8          1.9         0.4     setosa  45
#> 46           4.8         3.0          1.4         0.3     setosa  46
#> 47           5.1         3.8          1.6         0.2     setosa  47
#> 48           4.6         3.2          1.4         0.2     setosa  48
#> 49           5.3         3.7          1.5         0.2     setosa  49
#> 50           5.0         3.3          1.4         0.2     setosa  50
#> 51           7.0         3.2          4.7         1.4 versicolor  51
#> 52           6.4         3.2          4.5         1.5 versicolor  52
#> 53           6.9         3.1          4.9         1.5 versicolor  53
#> 54           5.5         2.3          4.0         1.3 versicolor  54
#> 55           6.5         2.8          4.6         1.5 versicolor  55
#> 56           5.7         2.8          4.5         1.3 versicolor  56
#> 57           6.3         3.3          4.7         1.6 versicolor  57
#> 58           4.9         2.4          3.3         1.0 versicolor  58
#> 59           6.6         2.9          4.6         1.3 versicolor  59
#> 60           5.2         2.7          3.9         1.4 versicolor  60
#> 61           5.0         2.0          3.5         1.0 versicolor  61
#> 62           5.9         3.0          4.2         1.5 versicolor  62
#> 63           6.0         2.2          4.0         1.0 versicolor  63
#> 64           6.1         2.9          4.7         1.4 versicolor  64
#> 65           5.6         2.9          3.6         1.3 versicolor  65
#> 66           6.7         3.1          4.4         1.4 versicolor  66
#> 67           5.6         3.0          4.5         1.5 versicolor  67
#> 68           5.8         2.7          4.1         1.0 versicolor  68
#> 69           6.2         2.2          4.5         1.5 versicolor  69
#> 70           5.6         2.5          3.9         1.1 versicolor  70
#> 71           5.9         3.2          4.8         1.8 versicolor  71
#> 72           6.1         2.8          4.0         1.3 versicolor  72
#> 73           6.3         2.5          4.9         1.5 versicolor  73
#> 74           6.1         2.8          4.7         1.2 versicolor  74
#> 75           6.4         2.9          4.3         1.3 versicolor  75
#> 76           6.6         3.0          4.4         1.4 versicolor  76
#> 77           6.8         2.8          4.8         1.4 versicolor  77
#> 78           6.7         3.0          5.0         1.7 versicolor  78
#> 79           6.0         2.9          4.5         1.5 versicolor  79
#> 80           5.7         2.6          3.5         1.0 versicolor  80
#> 81           5.5         2.4          3.8         1.1 versicolor  81
#> 82           5.5         2.4          3.7         1.0 versicolor  82
#> 83           5.8         2.7          3.9         1.2 versicolor  83
#> 84           6.0         2.7          5.1         1.6 versicolor  84
#> 85           5.4         3.0          4.5         1.5 versicolor  85
#> 86           6.0         3.4          4.5         1.6 versicolor  86
#> 87           6.7         3.1          4.7         1.5 versicolor  87
#> 88           6.3         2.3          4.4         1.3 versicolor  88
#> 89           5.6         3.0          4.1         1.3 versicolor  89
#> 90           5.5         2.5          4.0         1.3 versicolor  90
#> 91           5.5         2.6          4.4         1.2 versicolor  91
#> 92           6.1         3.0          4.6         1.4 versicolor  92
#> 93           5.8         2.6          4.0         1.2 versicolor  93
#> 94           5.0         2.3          3.3         1.0 versicolor  94
#> 95           5.6         2.7          4.2         1.3 versicolor  95
#> 96           5.7         3.0          4.2         1.2 versicolor  96
#> 97           5.7         2.9          4.2         1.3 versicolor  97
#> 98           6.2         2.9          4.3         1.3 versicolor  98
#> 99           5.1         2.5          3.0         1.1 versicolor  99
#> 100          5.7         2.8          4.1         1.3 versicolor 100
#> 101          6.3         3.3          6.0         2.5  virginica 101
#> 102          5.8         2.7          5.1         1.9  virginica 102
#> 103          7.1         3.0          5.9         2.1  virginica 103
#> 104          6.3         2.9          5.6         1.8  virginica 104
#> 105          6.5         3.0          5.8         2.2  virginica 105
#> 106          7.6         3.0          6.6         2.1  virginica 106
#> 107          4.9         2.5          4.5         1.7  virginica 107
#> 108          7.3         2.9          6.3         1.8  virginica 108
#> 109          6.7         2.5          5.8         1.8  virginica 109
#> 110          7.2         3.6          6.1         2.5  virginica 110
#> 111          6.5         3.2          5.1         2.0  virginica 111
#> 112          6.4         2.7          5.3         1.9  virginica 112
#> 113          6.8         3.0          5.5         2.1  virginica 113
#> 114          5.7         2.5          5.0         2.0  virginica 114
#> 115          5.8         2.8          5.1         2.4  virginica 115
#> 116          6.4         3.2          5.3         2.3  virginica 116
#> 117          6.5         3.0          5.5         1.8  virginica 117
#> 118          7.7         3.8          6.7         2.2  virginica 118
#> 119          7.7         2.6          6.9         2.3  virginica 119
#> 120          6.0         2.2          5.0         1.5  virginica 120
#> 121          6.9         3.2          5.7         2.3  virginica 121
#> 122          5.6         2.8          4.9         2.0  virginica 122
#> 123          7.7         2.8          6.7         2.0  virginica 123
#> 124          6.3         2.7          4.9         1.8  virginica 124
#> 125          6.7         3.3          5.7         2.1  virginica 125
#> 126          7.2         3.2          6.0         1.8  virginica 126
#> 127          6.2         2.8          4.8         1.8  virginica 127
#> 128          6.1         3.0          4.9         1.8  virginica 128
#> 129          6.4         2.8          5.6         2.1  virginica 129
#> 130          7.2         3.0          5.8         1.6  virginica 130
#> 131          7.4         2.8          6.1         1.9  virginica 131
#> 132          7.9         3.8          6.4         2.0  virginica 132
#> 133          6.4         2.8          5.6         2.2  virginica 133
#> 134          6.3         2.8          5.1         1.5  virginica 134
#> 135          6.1         2.6          5.6         1.4  virginica 135
#> 136          7.7         3.0          6.1         2.3  virginica 136
#> 137          6.3         3.4          5.6         2.4  virginica 137
#> 138          6.4         3.1          5.5         1.8  virginica 138
#> 139          6.0         3.0          4.8         1.8  virginica 139
#> 140          6.9         3.1          5.4         2.1  virginica 140
#> 141          6.7         3.1          5.6         2.4  virginica 141
#> 142          6.9         3.1          5.1         2.3  virginica 142
#> 143          5.8         2.7          5.1         1.9  virginica 143
#> 144          6.8         3.2          5.9         2.3  virginica 144
#> 145          6.7         3.3          5.7         2.5  virginica 145
#> 146          6.7         3.0          5.2         2.3  virginica 146
#> 147          6.3         2.5          5.0         1.9  virginica 147
#> 148          6.5         3.0          5.2         2.0  virginica 148
#> 149          6.2         3.4          5.4         2.3  virginica 149
#> 150          5.9         3.0          5.1         1.8  virginica 150
#>     sample_indicator sampling_prob
#> 1                  0            NA
#> 2                  0            NA
#> 3                  0            NA
#> 4                  0            NA
#> 5                  0            NA
#> 6                  0            NA
#> 7                  1           0.1
#> 8                  0            NA
#> 9                  0            NA
#> 10                 1           0.1
#> 11                 0            NA
#> 12                 0            NA
#> 13                 0            NA
#> 14                 0            NA
#> 15                 0            NA
#> 16                 0            NA
#> 17                 0            NA
#> 18                 0            NA
#> 19                 0            NA
#> 20                 0            NA
#> 21                 0            NA
#> 22                 0            NA
#> 23                 0            NA
#> 24                 0            NA
#> 25                 0            NA
#> 26                 0            NA
#> 27                 0            NA
#> 28                 0            NA
#> 29                 0            NA
#> 30                 1           0.1
#> 31                 0            NA
#> 32                 1           0.1
#> 33                 0            NA
#> 34                 0            NA
#> 35                 0            NA
#> 36                 0            NA
#> 37                 0            NA
#> 38                 0            NA
#> 39                 0            NA
#> 40                 0            NA
#> 41                 0            NA
#> 42                 0            NA
#> 43                 0            NA
#> 44                 0            NA
#> 45                 1           0.1
#> 46                 0            NA
#> 47                 0            NA
#> 48                 0            NA
#> 49                 0            NA
#> 50                 0            NA
#> 51                 1           0.1
#> 52                 1           0.1
#> 53                 0            NA
#> 54                 1           0.1
#> 55                 0            NA
#> 56                 0            NA
#> 57                 0            NA
#> 58                 0            NA
#> 59                 0            NA
#> 60                 0            NA
#> 61                 0            NA
#> 62                 0            NA
#> 63                 0            NA
#> 64                 0            NA
#> 65                 1           0.1
#> 66                 0            NA
#> 67                 1           0.1
#> 68                 0            NA
#> 69                 0            NA
#> 70                 0            NA
#> 71                 0            NA
#> 72                 0            NA
#> 73                 0            NA
#> 74                 0            NA
#> 75                 0            NA
#> 76                 0            NA
#> 77                 0            NA
#> 78                 0            NA
#> 79                 0            NA
#> 80                 0            NA
#> 81                 0            NA
#> 82                 0            NA
#> 83                 0            NA
#> 84                 0            NA
#> 85                 0            NA
#> 86                 0            NA
#> 87                 0            NA
#> 88                 0            NA
#> 89                 0            NA
#> 90                 0            NA
#> 91                 0            NA
#> 92                 0            NA
#> 93                 0            NA
#> 94                 0            NA
#> 95                 0            NA
#> 96                 0            NA
#> 97                 0            NA
#> 98                 0            NA
#> 99                 0            NA
#> 100                0            NA
#> 101                0            NA
#> 102                1           0.1
#> 103                1           0.1
#> 104                0            NA
#> 105                0            NA
#> 106                0            NA
#> 107                0            NA
#> 108                0            NA
#> 109                0            NA
#> 110                0            NA
#> 111                0            NA
#> 112                0            NA
#> 113                0            NA
#> 114                0            NA
#> 115                0            NA
#> 116                0            NA
#> 117                0            NA
#> 118                0            NA
#> 119                0            NA
#> 120                0            NA
#> 121                0            NA
#> 122                0            NA
#> 123                0            NA
#> 124                0            NA
#> 125                0            NA
#> 126                0            NA
#> 127                0            NA
#> 128                0            NA
#> 129                0            NA
#> 130                0            NA
#> 131                0            NA
#> 132                0            NA
#> 133                0            NA
#> 134                0            NA
#> 135                0            NA
#> 136                0            NA
#> 137                1           0.1
#> 138                0            NA
#> 139                0            NA
#> 140                0            NA
#> 141                1           0.1
#> 142                0            NA
#> 143                1           0.1
#> 144                0            NA
#> 145                0            NA
#> 146                0            NA
#> 147                0            NA
#> 148                0            NA
#> 149                0            NA
#> 150                0            NA

# If some units had already been sampled
iris$already_sampled <- rbinom(nrow(iris), 1, 0.25)

sample_strata(
  data = iris, strata = "Species", id = "id",
  already_sampled = "already_sampled",
  design_data = design, design_strata = "strata",
  n_allocated = "n_to_sample"
)
#>     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species  id
#> 1            5.1         3.5          1.4         0.2     setosa   1
#> 2            4.9         3.0          1.4         0.2     setosa   2
#> 3            4.7         3.2          1.3         0.2     setosa   3
#> 4            4.6         3.1          1.5         0.2     setosa   4
#> 5            5.0         3.6          1.4         0.2     setosa   5
#> 6            5.4         3.9          1.7         0.4     setosa   6
#> 7            4.6         3.4          1.4         0.3     setosa   7
#> 8            5.0         3.4          1.5         0.2     setosa   8
#> 9            4.4         2.9          1.4         0.2     setosa   9
#> 10           4.9         3.1          1.5         0.1     setosa  10
#> 11           5.4         3.7          1.5         0.2     setosa  11
#> 12           4.8         3.4          1.6         0.2     setosa  12
#> 13           4.8         3.0          1.4         0.1     setosa  13
#> 14           4.3         3.0          1.1         0.1     setosa  14
#> 15           5.8         4.0          1.2         0.2     setosa  15
#> 16           5.7         4.4          1.5         0.4     setosa  16
#> 17           5.4         3.9          1.3         0.4     setosa  17
#> 18           5.1         3.5          1.4         0.3     setosa  18
#> 19           5.7         3.8          1.7         0.3     setosa  19
#> 20           5.1         3.8          1.5         0.3     setosa  20
#> 21           5.4         3.4          1.7         0.2     setosa  21
#> 22           5.1         3.7          1.5         0.4     setosa  22
#> 23           4.6         3.6          1.0         0.2     setosa  23
#> 24           5.1         3.3          1.7         0.5     setosa  24
#> 25           4.8         3.4          1.9         0.2     setosa  25
#> 26           5.0         3.0          1.6         0.2     setosa  26
#> 27           5.0         3.4          1.6         0.4     setosa  27
#> 28           5.2         3.5          1.5         0.2     setosa  28
#> 29           5.2         3.4          1.4         0.2     setosa  29
#> 30           4.7         3.2          1.6         0.2     setosa  30
#> 31           4.8         3.1          1.6         0.2     setosa  31
#> 32           5.4         3.4          1.5         0.4     setosa  32
#> 33           5.2         4.1          1.5         0.1     setosa  33
#> 34           5.5         4.2          1.4         0.2     setosa  34
#> 35           4.9         3.1          1.5         0.2     setosa  35
#> 36           5.0         3.2          1.2         0.2     setosa  36
#> 37           5.5         3.5          1.3         0.2     setosa  37
#> 38           4.9         3.6          1.4         0.1     setosa  38
#> 39           4.4         3.0          1.3         0.2     setosa  39
#> 40           5.1         3.4          1.5         0.2     setosa  40
#> 41           5.0         3.5          1.3         0.3     setosa  41
#> 42           4.5         2.3          1.3         0.3     setosa  42
#> 43           4.4         3.2          1.3         0.2     setosa  43
#> 44           5.0         3.5          1.6         0.6     setosa  44
#> 45           5.1         3.8          1.9         0.4     setosa  45
#> 46           4.8         3.0          1.4         0.3     setosa  46
#> 47           5.1         3.8          1.6         0.2     setosa  47
#> 48           4.6         3.2          1.4         0.2     setosa  48
#> 49           5.3         3.7          1.5         0.2     setosa  49
#> 50           5.0         3.3          1.4         0.2     setosa  50
#> 51           7.0         3.2          4.7         1.4 versicolor  51
#> 52           6.4         3.2          4.5         1.5 versicolor  52
#> 53           6.9         3.1          4.9         1.5 versicolor  53
#> 54           5.5         2.3          4.0         1.3 versicolor  54
#> 55           6.5         2.8          4.6         1.5 versicolor  55
#> 56           5.7         2.8          4.5         1.3 versicolor  56
#> 57           6.3         3.3          4.7         1.6 versicolor  57
#> 58           4.9         2.4          3.3         1.0 versicolor  58
#> 59           6.6         2.9          4.6         1.3 versicolor  59
#> 60           5.2         2.7          3.9         1.4 versicolor  60
#> 61           5.0         2.0          3.5         1.0 versicolor  61
#> 62           5.9         3.0          4.2         1.5 versicolor  62
#> 63           6.0         2.2          4.0         1.0 versicolor  63
#> 64           6.1         2.9          4.7         1.4 versicolor  64
#> 65           5.6         2.9          3.6         1.3 versicolor  65
#> 66           6.7         3.1          4.4         1.4 versicolor  66
#> 67           5.6         3.0          4.5         1.5 versicolor  67
#> 68           5.8         2.7          4.1         1.0 versicolor  68
#> 69           6.2         2.2          4.5         1.5 versicolor  69
#> 70           5.6         2.5          3.9         1.1 versicolor  70
#> 71           5.9         3.2          4.8         1.8 versicolor  71
#> 72           6.1         2.8          4.0         1.3 versicolor  72
#> 73           6.3         2.5          4.9         1.5 versicolor  73
#> 74           6.1         2.8          4.7         1.2 versicolor  74
#> 75           6.4         2.9          4.3         1.3 versicolor  75
#> 76           6.6         3.0          4.4         1.4 versicolor  76
#> 77           6.8         2.8          4.8         1.4 versicolor  77
#> 78           6.7         3.0          5.0         1.7 versicolor  78
#> 79           6.0         2.9          4.5         1.5 versicolor  79
#> 80           5.7         2.6          3.5         1.0 versicolor  80
#> 81           5.5         2.4          3.8         1.1 versicolor  81
#> 82           5.5         2.4          3.7         1.0 versicolor  82
#> 83           5.8         2.7          3.9         1.2 versicolor  83
#> 84           6.0         2.7          5.1         1.6 versicolor  84
#> 85           5.4         3.0          4.5         1.5 versicolor  85
#> 86           6.0         3.4          4.5         1.6 versicolor  86
#> 87           6.7         3.1          4.7         1.5 versicolor  87
#> 88           6.3         2.3          4.4         1.3 versicolor  88
#> 89           5.6         3.0          4.1         1.3 versicolor  89
#> 90           5.5         2.5          4.0         1.3 versicolor  90
#> 91           5.5         2.6          4.4         1.2 versicolor  91
#> 92           6.1         3.0          4.6         1.4 versicolor  92
#> 93           5.8         2.6          4.0         1.2 versicolor  93
#> 94           5.0         2.3          3.3         1.0 versicolor  94
#> 95           5.6         2.7          4.2         1.3 versicolor  95
#> 96           5.7         3.0          4.2         1.2 versicolor  96
#> 97           5.7         2.9          4.2         1.3 versicolor  97
#> 98           6.2         2.9          4.3         1.3 versicolor  98
#> 99           5.1         2.5          3.0         1.1 versicolor  99
#> 100          5.7         2.8          4.1         1.3 versicolor 100
#> 101          6.3         3.3          6.0         2.5  virginica 101
#> 102          5.8         2.7          5.1         1.9  virginica 102
#> 103          7.1         3.0          5.9         2.1  virginica 103
#> 104          6.3         2.9          5.6         1.8  virginica 104
#> 105          6.5         3.0          5.8         2.2  virginica 105
#> 106          7.6         3.0          6.6         2.1  virginica 106
#> 107          4.9         2.5          4.5         1.7  virginica 107
#> 108          7.3         2.9          6.3         1.8  virginica 108
#> 109          6.7         2.5          5.8         1.8  virginica 109
#> 110          7.2         3.6          6.1         2.5  virginica 110
#> 111          6.5         3.2          5.1         2.0  virginica 111
#> 112          6.4         2.7          5.3         1.9  virginica 112
#> 113          6.8         3.0          5.5         2.1  virginica 113
#> 114          5.7         2.5          5.0         2.0  virginica 114
#> 115          5.8         2.8          5.1         2.4  virginica 115
#> 116          6.4         3.2          5.3         2.3  virginica 116
#> 117          6.5         3.0          5.5         1.8  virginica 117
#> 118          7.7         3.8          6.7         2.2  virginica 118
#> 119          7.7         2.6          6.9         2.3  virginica 119
#> 120          6.0         2.2          5.0         1.5  virginica 120
#> 121          6.9         3.2          5.7         2.3  virginica 121
#> 122          5.6         2.8          4.9         2.0  virginica 122
#> 123          7.7         2.8          6.7         2.0  virginica 123
#> 124          6.3         2.7          4.9         1.8  virginica 124
#> 125          6.7         3.3          5.7         2.1  virginica 125
#> 126          7.2         3.2          6.0         1.8  virginica 126
#> 127          6.2         2.8          4.8         1.8  virginica 127
#> 128          6.1         3.0          4.9         1.8  virginica 128
#> 129          6.4         2.8          5.6         2.1  virginica 129
#> 130          7.2         3.0          5.8         1.6  virginica 130
#> 131          7.4         2.8          6.1         1.9  virginica 131
#> 132          7.9         3.8          6.4         2.0  virginica 132
#> 133          6.4         2.8          5.6         2.2  virginica 133
#> 134          6.3         2.8          5.1         1.5  virginica 134
#> 135          6.1         2.6          5.6         1.4  virginica 135
#> 136          7.7         3.0          6.1         2.3  virginica 136
#> 137          6.3         3.4          5.6         2.4  virginica 137
#> 138          6.4         3.1          5.5         1.8  virginica 138
#> 139          6.0         3.0          4.8         1.8  virginica 139
#> 140          6.9         3.1          5.4         2.1  virginica 140
#> 141          6.7         3.1          5.6         2.4  virginica 141
#> 142          6.9         3.1          5.1         2.3  virginica 142
#> 143          5.8         2.7          5.1         1.9  virginica 143
#> 144          6.8         3.2          5.9         2.3  virginica 144
#> 145          6.7         3.3          5.7         2.5  virginica 145
#> 146          6.7         3.0          5.2         2.3  virginica 146
#> 147          6.3         2.5          5.0         1.9  virginica 147
#> 148          6.5         3.0          5.2         2.0  virginica 148
#> 149          6.2         3.4          5.4         2.3  virginica 149
#> 150          5.9         3.0          5.1         1.8  virginica 150
#>     already_sampled sample_indicator
#> 1                 0                0
#> 2                 0                0
#> 3                 0                0
#> 4                 0                0
#> 5                 1                0
#> 6                 0                0
#> 7                 1                0
#> 8                 1                0
#> 9                 0                0
#> 10                0                0
#> 11                0                0
#> 12                0                0
#> 13                1                0
#> 14                1                0
#> 15                0                0
#> 16                0                1
#> 17                0                0
#> 18                0                1
#> 19                0                0
#> 20                0                1
#> 21                0                0
#> 22                0                0
#> 23                0                1
#> 24                1                0
#> 25                1                0
#> 26                1                0
#> 27                0                0
#> 28                0                0
#> 29                1                0
#> 30                1                0
#> 31                0                0
#> 32                0                0
#> 33                1                0
#> 34                0                0
#> 35                1                0
#> 36                1                0
#> 37                0                0
#> 38                0                0
#> 39                1                0
#> 40                1                0
#> 41                0                1
#> 42                0                0
#> 43                1                0
#> 44                0                0
#> 45                0                0
#> 46                0                0
#> 47                0                0
#> 48                0                0
#> 49                0                0
#> 50                0                0
#> 51                0                0
#> 52                0                0
#> 53                0                0
#> 54                0                0
#> 55                0                0
#> 56                1                0
#> 57                0                0
#> 58                0                0
#> 59                0                0
#> 60                0                0
#> 61                0                0
#> 62                1                0
#> 63                0                0
#> 64                0                0
#> 65                1                0
#> 66                0                0
#> 67                0                0
#> 68                0                0
#> 69                0                1
#> 70                0                0
#> 71                0                1
#> 72                0                0
#> 73                0                0
#> 74                0                0
#> 75                0                0
#> 76                0                0
#> 77                1                0
#> 78                0                0
#> 79                0                0
#> 80                0                0
#> 81                1                0
#> 82                0                1
#> 83                0                0
#> 84                0                0
#> 85                0                0
#> 86                0                0
#> 87                0                0
#> 88                0                0
#> 89                0                0
#> 90                0                1
#> 91                0                0
#> 92                0                0
#> 93                0                0
#> 94                0                0
#> 95                0                0
#> 96                0                0
#> 97                0                0
#> 98                0                0
#> 99                0                0
#> 100               0                1
#> 101               0                0
#> 102               0                0
#> 103               0                0
#> 104               0                0
#> 105               0                0
#> 106               0                0
#> 107               0                1
#> 108               0                0
#> 109               0                0
#> 110               0                0
#> 111               0                0
#> 112               0                0
#> 113               0                1
#> 114               0                0
#> 115               0                0
#> 116               0                0
#> 117               1                0
#> 118               0                0
#> 119               0                0
#> 120               0                0
#> 121               0                0
#> 122               1                0
#> 123               0                0
#> 124               1                0
#> 125               0                0
#> 126               0                1
#> 127               0                1
#> 128               1                0
#> 129               0                0
#> 130               0                0
#> 131               0                0
#> 132               0                0
#> 133               0                0
#> 134               0                0
#> 135               1                0
#> 136               0                0
#> 137               0                0
#> 138               0                0
#> 139               0                1
#> 140               0                0
#> 141               0                0
#> 142               0                0
#> 143               1                0
#> 144               1                0
#> 145               0                0
#> 146               0                0
#> 147               0                0
#> 148               1                0
#> 149               0                0
#> 150               0                0